+++ src/gorilla_notes/components.cljs
@@ -2,8 +2,10 @@
   (:require [gorilla-notes.components.code :as code]
             [gorilla-notes.components.input :as input]
             [gorilla-notes.components.dataset :as dataset]
+            [gorilla-notes.components.leaflet :as leaflet]
             [pinkie.pinkie :refer-macros [register-component]]))
 
-(register-component :p/code code/Code)
-(register-component :p/slider input/Slider)
-(register-component :p/dataset dataset/Dataset)
+(register-component :p/code code/code-block)
+(register-component :p/slider input/slider)
+(register-component :p/dataset dataset/dataset-view)
+(register-component :p/leaflet leaflet/leaflet-map)