+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file. This change
 ## [0.5.0-SNAPSHOT]
 - updated core.async dep version
 - took care of ag-grid css
+- added leaflet component (a bit more flexible than the Gorilla-UI one)
+- some internal renaming
 
 ## [0.4.0] - 2020-09-21
 - changes in aesthetics, configurability